Company tags
This PR adds resolvers to create and list tags, and allows to assign tags to companies. Specifically:

New resolver list-tags takes some criteria (for now just a tag type) and returns all tags with that type.
New resolver create-tag takes a label and type, generates an ID and a slug, and creates and returns a tag. If the tag with this slug already existed, it returns the pre-existing one instead.
The Company object now has a tags field resolver that returns all associated tags.
Update company input now includes tag ids. The resolver will check whether they actually correspond to one of the existing tags.
This is all server-side, so cant